# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MOSTLYVERS API
  version: 1.0.0
  description: Contract for the Android reader and Owner dashboard. FREE_LAUNCH grants non-financial entitlements; payment routes remain reserved and fail closed.
servers:
  - url: http://localhost:5001/v1
  - url: https://api.example.com/v1
...
paths:
  /auth/signup:
    post: { tags: [Authentication], operationId: readerSignup, requestBody: { required: true, content: { multipart/form-data: { schema: { type: object, required: [name, age, gender, email, phone, password, device, otpChallengeId, otp], properties: { name: {type: string}, age: {type: integer}, gender: {type: string}, email: {type: string, format: email}, phone: {type: string}, password: {type: string, format: password}, otpChallengeId: {type: string, format: uuid}, otp: {type: string, pattern: '^[0-9]{6}$'}, device: {type: string, description: JSON-encoded DeviceContext}, profilePicture: {type: string, format: binary} } } } } }, responses: { '201': { $ref: '#/components/responses/ReaderSession' }, default: { $ref: '#/components/responses/Error' } } }
...
  /auth/login:
    post: { tags: [Authentication], operationId: readerLogin, requestBody: { $ref: '#/components/requestBodies/ReaderLogin' }, responses: { '200': { $ref: '#/components/responses/ReaderSession' }, default: { $ref: '#/components/responses/Error' } } }
  /auth/refresh:
    post: { tags: [Authentication], operationId: readerRefresh, responses: { '200': { $ref: '#/components/responses/ReaderSession' }, default: { $ref: '#/components/responses/Error' } } }
...
  /me/device-transfer/free: { post: { tags: [Reader], operationId: transferDeviceFree, security: [ { bearerAuth: [] } ], responses: { '200': { $ref: '#/components/responses/Object' }, default: { $ref: '#/components/responses/Error' } } } }
...
  /admin/auth/login: { post: { tags: [Authentication], operationId: adminLogin, responses: { '200': { $ref: '#/components/responses/AdminSession' }, default: { $ref: '#/components/responses/Error' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    internalToken: { type: http, scheme: bearer }
...
  requestBodies:
    ReaderLogin: { required: true, content: { application/json: { schema: { type: object, required: [email, password, device], properties: { email: {type: string, format: email}, password: {type: string, format: password}, device: {$ref: '#/components/schemas/DeviceContext'} } } } } }
  responses:
    Object: { description: Success, content: { application/json: { schema: {type: object, additionalProperties: true} } } }
...
    ReaderSession: { description: Reader session, content: { application/json: { schema: {$ref: '#/components/schemas/ReaderSession'} } } }
    AdminSession: { description: Owner session, content: { application/json: { schema: {type: object, additionalProperties: true} } } }
    Error: { description: API error, content: { application/json: { schema: {$ref: '#/components/schemas/ErrorEnvelope'} } } }
...
  schemas:
...
    DeviceContext: { type: object, required: [installationId, displayName, platform, osVersion, appVersion, packageName], properties: {installationId: {type: string}, displayName: {type: string}, platform: {type: string, const: ANDROID}, osVersion: {type: string}, appVersion: {type: string}, packageName: {type: string}, integrityToken: {type: string}, integrityRequestHash: {type: string}} }
    ReaderSession: { type: object, required: [accessToken, accessTokenExpiresAt, refreshToken, reader, accessMode, deviceAccess], properties: {accessToken: {type: string}, accessTokenExpiresAt: {type: string, format: date-time}, refreshToken: {type: string}, reader: {type: object, additionalProperties: true}, accessMode: {enum: [FULL, BROWSE_ONLY]}, deviceAccess: {type: object, additionalProperties: true}} }
...
    ErrorEnvelope: { type: object, required: [error], properties: {error: {type: object, required: [code, message, requestId], properties: {code: {type: string}, message: {type: string}, requestId: {type: string}, details: {type: object, additionalProperties: true}}}} }